+++ curve25519-dalek/src/field.rs
@@ -25,8 +25,6 @@
 
 #![allow(unused_qualifications)]
 
-use core::cmp::{Eq, PartialEq};
-
 use cfg_if::cfg_if;
 
 use subtle::Choice;
@@ -39,11 +37,6 @@ use crate::constants;
 
 cfg_if! {
     if #[cfg(curve25519_dalek_backend = "fiat")] {
-        #[cfg(curve25519_dalek_bits = "32")]
-        pub use backend::serial::fiat_u32::field::*;
-        #[cfg(curve25519_dalek_bits = "64")]
-        pub use backend::serial::fiat_u64::field::*;
-
         /// A `FieldElement` represents an element of the field
         /// \\( \mathbb Z / (2\^{255} - 19)\\).
         ///
@@ -52,7 +45,7 @@ cfg_if! {
         ///
         /// Using formally-verified field arithmetic from fiat-crypto.
         #[cfg(curve25519_dalek_bits = "32")]
-        pub type FieldElement = backend::serial::fiat_u32::field::FieldElement2625;
+        pub(crate) type FieldElement = backend::serial::fiat_u32::field::FieldElement2625;
 
         /// A `FieldElement` represents an element of the field
         /// \\( \mathbb Z / (2\^{255} - 19)\\).
@@ -62,25 +55,21 @@ cfg_if! {
         ///
         /// Using formally-verified field arithmetic from fiat-crypto.
         #[cfg(curve25519_dalek_bits = "64")]
-        pub type FieldElement = backend::serial::fiat_u64::field::FieldElement51;
+        pub(crate) type FieldElement = backend::serial::fiat_u64::field::FieldElement51;
     } else if #[cfg(curve25519_dalek_bits = "64")] {
-        pub use crate::backend::serial::u64::field::*;
-
         /// A `FieldElement` represents an element of the field
         /// \\( \mathbb Z / (2\^{255} - 19)\\).
         ///
         /// The `FieldElement` type is an alias for one of the platform-specific
         /// implementations.
-        pub type FieldElement = backend::serial::u64::field::FieldElement51;
+        pub(crate) type FieldElement = backend::serial::u64::field::FieldElement51;
     } else {
-        pub use backend::serial::u32::field::*;
-
         /// A `FieldElement` represents an element of the field
         /// \\( \mathbb Z / (2\^{255} - 19)\\).
         ///
         /// The `FieldElement` type is an alias for one of the platform-specific
         /// implementations.
-        pub type FieldElement = backend::serial::u32::field::FieldElement2625;
+        pub(crate) type FieldElement = backend::serial::u32::field::FieldElement2625;
     }
 }
 
@@ -109,7 +98,7 @@ impl FieldElement {
     /// # Return
     ///
     /// If negative, return `Choice(1)`.  Otherwise, return `Choice(0)`.
-    pub fn is_negative(&self) -> Choice {
+    pub(crate) fn is_negative(&self) -> Choice {
         let bytes = self.as_bytes();
         (bytes[0] & 1).into()
     }
@@ -119,7 +108,7 @@ impl FieldElement {
     /// # Return
     ///
     /// If zero, return `Choice(1)`.  Otherwise, return `Choice(0)`.
-    pub fn is_zero(&self) -> Choice {
+    pub(crate) fn is_zero(&self) -> Choice {
         let zero = [0u8; 32];
         let bytes = self.as_bytes();
 
@@ -165,11 +154,11 @@ impl FieldElement {
         (t19, t3)
     }
 
-    /// Given a slice of public `FieldElements`, replace each with its inverse.
+    /// Given a slice of pub(crate)lic `FieldElements`, replace each with its inverse.
     ///
     /// When an input `FieldElement` is zero, its value is unchanged.
     #[cfg(feature = "alloc")]
-    pub fn batch_invert(inputs: &mut [FieldElement]) {
+    pub(crate) fn batch_invert(inputs: &mut [FieldElement]) {
         // Montgomery’s Trick and Fast Implementation of Masked AES
         // Genelle, Prouff and Quisquater
         // Section 3.2
@@ -214,7 +203,7 @@ impl FieldElement {
     /// This function returns zero on input zero.
     #[rustfmt::skip] // keep alignment of explanatory comments
     #[allow(clippy::let_and_return)]
-    pub fn invert(&self) -> FieldElement {
+    pub(crate) fn invert(&self) -> FieldElement {
         // The bits of p-2 = 2^255 -19 -2 are 11010111111...11.
         //
         //                                 nonzero bits of exponent
@@ -251,7 +240,7 @@ impl FieldElement {
     /// - `(Choice(0), zero)        ` if `v` is zero and `u` is nonzero;
     /// - `(Choice(0), +sqrt(i*u/v))` if `u/v` is nonsquare (so `i*u/v` is square).
     ///
-    pub fn sqrt_ratio_i(u: &FieldElement, v: &FieldElement) -> (Choice, FieldElement) {
+    pub(crate) fn sqrt_ratio_i(u: &FieldElement, v: &FieldElement) -> (Choice, FieldElement) {
         // Using the same trick as in ed25519 decoding, we merge the
         // inversion, the square root, and the square test as follows.
         //
@@ -311,7 +300,7 @@ impl FieldElement {
     /// - `(Choice(0), zero)           ` if `self` is zero;
     /// - `(Choice(0), +sqrt(i/self))  ` if `self` is a nonzero nonsquare;
     ///
-    pub fn invsqrt(&self) -> (Choice, FieldElement) {
+    pub(crate) fn invsqrt(&self) -> (Choice, FieldElement) {
         FieldElement::sqrt_ratio_i(&FieldElement::ONE, self)
     }
 }
@@ -319,7 +308,6 @@ impl FieldElement {
 #[cfg(test)]
 mod test {
     use crate::field::*;
-    use subtle::ConditionallyNegatable;
 
     /// Random element a of GF(2^255-19), from Sage
     /// a = 1070314506888354081329385823235218444233221\