+++ ed25519-dalek/tests/validation_criteria.rs
@@ -22,7 +22,7 @@ const VERIFY_ALLOWED_EDGECASES: &[Flag] = &[
 const VERIFY_STRICT_ALLOWED_EDGECASES: &[Flag] =
     &[Flag::LowOrderComponentA, Flag::LowOrderComponentR];
 
-/// Each variant describes a specfiic edge case that can occur in an Ed25519 signature. Refer to
+/// Each variant describes a specific edge case that can occur in an Ed25519 signature. Refer to
 /// the test vector [README][] for more info.
 ///
 /// [README]: https://github.com/C2SP/CCTV/blob/5ea85644bd035c555900a2f707f7e4c31ea65ced/ed25519vectors/README.md